tests/ioctl: adjust size of variable passed to ioctl(2)
FIOQSIZE returns a result of type loff_t, which maps to long long int,
so make sure that the variable is large enough to not cause stack
corruption.

Fixes: b11701a55614 ("tests/ioctl: switch to FIOQSIZE")

+++ tests/ioctl/test_ioctl.c
@@ -18,7 +18,7 @@ int main(int argc, char **argv)
 
 	int fd;
 	int rc;
-	int val = 0;
+	long long int val = 0;
 
 	fd = open(argv[1], O_RDONLY, 0);
